requirements.txt
torch
transformers
accelerate
peft
huggingface_hub
pillow
requests
runpod